Add sysfs PMU caching on initialization
In order to accommodate the growing number of PMUs active and to
handle hybrid processors better, this patch adds sysfs PMU perf_events
information caching to avoid going back to sysfs for each encoded event.
The caching stores the name of PMU, e.g., armv8_pmu3, and the perf_events
type which is then use to build the perf_events encoding.

Signed-off-by: Stephane Eranian <eranian@gmail.com>