Remove references to /sys/devices to remain compatible with upstream
The PMUs will not appear in /sys/devices for much longer.
The proper way to access PMU directories is via:
/sys/bus/event_source/devices/

Where each PMU has a symlink.

It should be noted that this alternate directory is not new. It has
been there all along. Therefore it is okay to remove all references
to /sys/devices.

Signed-off-by: Stephane Eranian <eranian@gmail.com>